+++ cmake/nfxSiliconSources.cmake
@@ -35,6 +35,7 @@ set(NFX_SILICON_SOURCES
     src/gate/sn74/LS04.cpp
     src/gate/sn74/LS08.cpp
     src/gate/sn74/LS74.cpp
+    src/gate/sn74/LS138.cpp
     src/gate/sn74/LS373.cpp
     src/gate/sn74/LS374.cpp